+++ R/parseFuns.R
@@ -37,7 +37,7 @@
 "d4GELU", "ELU", "dELU", "d2ELU", "d2aELU", "dELUa", "d2ELUa", 
 "softplus", "dsoftplus", "d2softplus", "d3softplus", "d4softplus", 
 "SELU", "dSELU", "lReLU", "dlReLU", "PReLU", "dPReLU", "d2PReLU", 
-"dPReLUa", "dPReLUa1", "Swish", "dSwish", "linCmt", "et_", "rnorm", 
+"dPReLUa", "dPReLUa1", "Swish", "dSwish", "linCmt", "et_", "etInf_", "rnorm", 
 "rxnorm", "rxbinom", "rbinom", "rxcauchy", "rcauchy", "rchisq", 
 "rxchisq", "rexp", "rxexp", "rbeta", "rxbeta", "rgeom", "rxgeom", 
 "rxpois", "rpois", "rxt", "rt")
@@ -81,4 +81,4 @@ d4GELU = 1, ELU = 2, dELU = 2, d2ELU = 2, d2aELU = 2, dELUa = 2,
 d2ELUa = 2, softplus = 1, dsoftplus = 1, d2softplus = 1, d3softplus = 1, 
 d4softplus = 1, SELU = 1, dSELU = 1, lReLU = 1, dlReLU = 1, PReLU = 2, 
 dPReLU = 2, d2PReLU = 2, dPReLUa = 2, dPReLUa1 = 2, Swish = 1, 
-dSwish = 1)
+dSwish = 1, etInf_ = 4)