+++ src/init.c
@@ -783,6 +783,8 @@ void R_init_rxode2(DllInfo *info){
   // C callable to assign environments.
   extern int _rxPushDosingEvent(double, double, int, void *);
   R_RegisterCCallable("rxode2", "_rxPushDosingEvent", (DL_FUNC) &_rxPushDosingEvent);
+  extern double _rxGetAMT(int, int, double, double, double *);
+  R_RegisterCCallable("rxode2", "_rxGetAMT", (DL_FUNC) &_rxGetAMT);
   R_RegisterCCallable("rxode2", "linCmtA", (DL_FUNC) &linCmtA);
   R_RegisterCCallable("rxode2", "linCmtB", (DL_FUNC) &linCmtB);
   R_RegisterCCallable("rxode2", "_rxode2_rxRmvnSEXP",